compra_producto

+++ app/Http/Controllers/CategoriaController.php
@@ -14,6 +14,9 @@ class CategoriaController extends Controller
      */
     public function index()
     {
+
+        $this->authorize('view', new Categoria() );
+
         $datos = Categoria::all();
         return view('categoria.index',compact('datos'));
     }
@@ -25,74 +28,54 @@ class CategoriaController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', new Categoria());
+
         return view('categoria.create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
+
+        $this->authorize('update', new Categoria());
+
         $datos = new Categoria();
         $datos->nombre_categoria =$request->nombre_cate;
-  
         $datos->save();
-      
+
         return redirect()->route('categoria.index');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Proveedor  $proveedor
-     * @return \Illuminate\Http\Response
-     */
     public function show($id)
     {
         $categoria=Categoria::findOrFail($id);
-        
         return view('categoria.show', compact('categoria'));
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Proveedor  $proveedor
-     * @return \Illuminate\Http\Response
-     */
     public function edit($id)
     {
+
+        $this->authorize('update', new Categoria());
+
         $categoria=Categoria::findOrFail($id);
         return view('categoria.edit', compact('categoria'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Proveedor  $proveedor
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request,$id)
     {
+
+        $this->authorize('update', new Categoria());
+
         $categoria=Categoria::findOrFail($id);
         $categoria->nombre_categoria =$request->nombre_cate;
-  
         $categoria->update();
         return redirect()->route('categoria.index');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Proveedor  $proveedor
-     * @return \Illuminate\Http\Response
-     */
     public function destroy($id)
     {
+
+        $this->authorize('delete', new Categoria());
+
         $categoria=Categoria::findOrFail($id);
         $categoria->delete();
         return redirect()->route('categoria.index');
